+++ icat/__init__.py
@@ -17,4 +17,4 @@ from icat.anchors import DictionaryAnchor, SimilarityFunctionAnchor, TFIDFAnchor
 # make the important things directly accessible off top level module
 from icat.model import Model
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"