+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Useful imports to top level module.
 * Activity indicators to anchors in anchor list table.
 * `add_anchor` function directly to model.
+* Saving/loading functionality to anchors and anchor lists.
 
 ### Fixed
 * Text fields not applying changes when user clicks away instead of hitting enter.